[Silver III] Title: 주유소, Time: 20 ms, Memory: 3588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -10,16 +10,16 @@ int main() {
 	int N;
 	cin >> N;
 
-	vector<int> load(N - 1);
-	vector<int> city(N);
+	vector<long long> load(N - 1);
+	vector<long long> city(N);
 
 	for (int i = 0; i < N - 1; i++)
 		cin >> load[i];
 	for (int i = 0; i < N; i++)
 		cin >> city[i];
 
-	int low_city = city[0];
-	int min = low_city * load[0];
+	long long low_city = city[0];
+	long long min = low_city * load[0];
 
 	for (int i = 1; i < N - 1; i++)
 	{
